rename string-cases to str and expose it on videojs
